Some quick branch cleanup, things into place to replace antiquated versions. Some brief fiddling with re-applying text processing, which I'm wrestling with how to implement in a way that isn't silly while still wanting it around for spot testing. Started implementation of a potential term <--> id conversion for improved memory effeciency.
