+++ uchicagoldr/textbatch.py
@@ -2,175 +2,166 @@ from collections import Counter
 from math import log
 
 from uchicagoldr.batch import Batch
-from uchicagoldr.textdocument import TextDocument
+from uchicagoldr.textitem import TextItem
 
 class TextBatch(Batch):
-    """
-    A subclass of the batch class, meant to potentially aid with text analysis
-    """
-    terms=[]
-    unique_terms=[]
-    term_counts=[]
-    doc_counts=Counter()
-    idfs={}
-    tf_idfs={}
-    batch_tf_idfs={}
-    vsm={}
+
+    term_index=None
+    doc_counts=None
+    idfs=None
+    iIdfs=None
+    tf_idfs=None
+    language_model=None
+    vsm=None
+    doc_vsms=None
+    term_map={}
 
     def __init__(self,path,root):
         Batch.__init__(self,root,path)
 
     def validate_items(self):
-        assert(len(self.get_items())>0)
         for item in self.get_items():
             try:
-                assert isinstance(item,TextDocument)
+                assert(isinstance(item,TextItem))
             except AssertionError:
                 return False
         return True
 
-    def find_terms(self,purge_raw=False):
-        assert(len(self.get_items())>0)
-        itemTerms=[]
+    def find_term_index(self):
+        index=Counter()
         for item in self.get_items():
-            assert isinstance(item,TextDocument)
-            item.set_raw_string(item.find_raw_string())
-            item.set_terms(item.find_terms(purge_raw=True))
-            itemTerms+=item.get_terms()
-        return itemTerms
+            index+=item.get_index()
+        return index
 
-    def set_terms(self,newTerms):
-        self.terms=newTerms
+    def get_term_index(self):
+        return self.term_index
 
-    def get_terms(self):
-        return self.terms
+    def set_term_index(self,newIndex):
+        assert(isinstance(newIndex,Counter))
+        self.term_index=newIndex
 
     def find_doc_counts(self):
-        assert(len(self.unique_terms)>0)
-        assert(len(self.get_items())>0)
         counts=Counter()
         for item in self.get_items():
-            item.set_terms(item.get_terms())
-            item.set_unique_terms(item.find_unique_terms())
-            for term in item.get_unique_terms():
+            for term in item.get_index():
                 counts[term]+=1
         return counts
-
-    def set_doc_counts(self,newCounts):
-        self.doc_counts=newCounts
 
     def get_doc_counts(self):
         return self.doc_counts
 
-    def find_unique_terms(self):
-        assert(len(self.get_terms())>0)
-        uniqueTerms=set(self.terms)
-        return uniqueTerms
-
-    def set_unique_terms(self,newUnique):
-        self.unique_terms=newUnique
-
-    def get_unique_terms(self):
-        return self.unique_terms
-
-    def find_term_counts(self):
-        assert(len(self.get_terms())>0)
-        assert(len(self.get_unique_terms())>0)
-        counts=Counter(self.get_terms())
-#        uniques=self.get_unique_terms()
-#        for term in uniques:
-#            counts[term]=self.terms.count(term)
-        return counts
-
-    def set_term_counts(self,newCounts):
-        self.term_counts=newCounts
-
-    def get_term_counts(self):
-        return self.term_counts
+    def set_doc_counts(self,newCounts):
+        assert(isinstance(newCounts,Counter))
+        self.doc_counts=newCounts
 
     def find_idfs(self):
-        termIDFS={}
-        for term in self.get_unique_terms():
+        termIDFs={}
+        for term in self.find_doc_counts():
             IDF=log(1+(len(self.get_items())/self.get_doc_counts()[term]))
-            termIDFS[term]=IDF
-        return termIDFS
+            termIDFs[term]=IDF
+        return termIDFs
 
     def get_idfs(self):
         return self.idfs
 
-    def set_idfs(self,newIDFS):
-        self.idfs=newIDFS
+    def set_idfs(self,newIDFs):
+        assert(isinstance(newIDFs,dict))
+        self.idfs=newIDFs
 
-    def find_item_tf_idfs(self):
-        assert(len(self.get_items())>0)
-        assert(len(self.get_unique_terms())>0)
-        assert(len(self.get_idfs())>0)
+    def find_iIdfs(self):
+        termIIDFs={}
+        for term in self.find_doc_counts():
+            IIDF=1+(self.get_doc_counts()[term]/len(self.get_items()))
+            termIIDFs[term]=IIDF
+        return termIIDFs
+
+    def get_iIdfs(self):
+        return self.iIdfs
+
+    def set_iIdfs(self,newIIDFS):
+        assert(isinstance(newIIDFS,dict))
+        self.iIdfs=newIIDFS
+
+    def find_tf_idfs(self):
         k=.5
-        itemTFIDFS={}
-        termIDFS=self.get_idfs()
+        itemTFIDFs={}
+        termIDFs=self.get_idfs()
         for item in self.get_items():
-            item.set_raw_string(item.find_raw_string())
-            item.set_terms(item.find_terms())
-            item.set_unique_terms(item.find_unique_terms())
-            item.set_term_counts(item.find_term_counts())
-            maxTerm=item.get_term_counts().most_common()[0][1]
-            tfidfs={}
-            for term in item.get_unique_terms():
-                if term in self.get_unique_terms():
-                    tf =k+(1-k)*(item.get_term_counts()[term]/maxTerm)
-                    idf=termIDFS[term]
-                    tfidfs[term]=tf*idf
-            itemTFIDFS[item.get_file_path()]=tfidfs
-        return itemTFIDFS
+            termTFIDFs={}
+            if len(item.get_index())==0:
+               itemTFIDFs[item.get_file_path()]=termTFIDFs 
+            else:
+                maxTerm=item.get_index().most_common()[0][1]
+                for term in item.get_index():
+                    tf=k+(1-k)*(item.get_index()[term]/maxTerm)
+                    idf=self.idfs[term]
+                    termTFIDFs[term]=tf*idf
+                itemTFIDFs[item.get_file_path()]=termTFIDFs
+        return itemTFIDFs
 
-    def get_item_tf_idfs(self):
+    def get_tf_idfs(self):
         return self.tf_idfs
 
-    def set_item_tf_idfs(self,newTFIDFs):
-        assert isinstance(newTFIDFs,dict)
+    def set_tf_idfs(self,newTFIDFs):
+        assert(isinstance(newTFIDFs,dict))
         self.tf_idfs=newTFIDFs
 
-    def find_batch_tf_idfs(self):
-        assert(len(self.get_term_counts())>0)
-        assert(len(self.get_unique_terms())>0)
-        assert(len(self.get_items())>0)
-        assert(len(self.get_term_counts())>0)
+    def find_language_model(self):
         k=.5
-        maxTerm=self.get_term_counts().most_common()[0][1]
-        batchTFIDFs=self.get_idfs()
-        for term in self.get_unique_terms():
-            IDF=batchTFIDFs[term]
-            tf=k+(1-k)*(self.get_term_counts()[term]/maxTerm)
-            tfidf=tf*IDF
-            batchTFIDFs[term]=tfidf
-        return batchTFIDFs
+        language_model={}
+        maxTerm=self.get_term_index().most_common()[0][1]
+        for term in self.get_term_index():
+            tf=k+(1-k)*(self.get_term_index()[term]/maxTerm)
+            iIdf=self.iIdfs[term]
+            language_model[term]=tf*iIdf
+        return language_model
 
-    def get_batch_tf_idfs(self):
-        return self.batch_tf_idfs
+    def get_language_model(self):
+        return self.language_model
 
-    def set_batch_tf_idfs(self,new_batch_tf_idfs):
-        assert(isinstance(new_batch_tf_idfs,dict))
-        self.batch_tf_idfs=new_batch_tf_idfs
+    def set_language_model(self,newlm):
+        assert(isinstance(newlm,dict))
+        self.language_model=newlm
+
+    def find_document_vector_space_models(self):
+        docVSMs={}
+        for item in self.get_items():
+            normalizedVectorLengths={}
+            squaredSum=0
+            for term in item.get_index():
+               squaredSum+=self.get_tf_idfs()[item.get_file_path()][term]**2
+            hypot=squaredSum**.5
+            for term in item.get_index():
+                normalizedVectorLengths[term]=self.get_tf_idfs()[item.get_file_path()][term]/hypot
+            docVSMs[item.get_file_path()]=normalizedVectorLengths
+        return docVSMs
+
+    def get_document_vector_space_models(self):
+        return self.doc_vsms
+
+    def set_document_vector_space_models(self,newDocVSMs):
+        assert(isinstance(newDocVSMs,dict))
+        self.doc_vsms=newDocVSMs
 
     def find_vector_space_model(self):
-        assert(len(self.get_batch_tf_idfs())>0)
         normalizedVectorLengths={}
-        edgeLength=0
-        for term in self.get_batch_tf_idfs():
-            edgeLength+=self.get_batch_tf_idfs()[term]**2
-        vectorLength=edgeLength**.5
-        for term in self.get_batch_tf_idfs():
-            normalizedVectorLengths[term]=self.get_batch_tf_idfs()[term]/vectorLength
+        squaredSum=0
+        for term in self.get_term_index():
+            squaredSum+=self.get_language_model()[term]**2
+        hypot=squaredSum**.5
+        for term in self.get_term_index():
+            normalizedVectorLengths[term]=self.get_language_model()[term]/hypot
         return normalizedVectorLengths
 
     def get_vector_space_model(self):
         return self.vsm
 
-    def set_vector_space_model(self,new_vsm):
-        self.vsm=new_vsm
+    def set_vector_space_model(self,newVSM):
+        assert(isinstance(newVSM,dict))
+        self.vsm=newVSM
 
     def find_similarity(self,other_vsm):
-        simVec={}
+       	simVec={}
         simPerc=0
         if len(self.get_vector_space_model())>len(other_vsm):
             for term in other_vsm:
